+++ app/services/ai/google-ai-service.ts
