+++ app/services/ai/open-ai-service.ts
